Move lib/library to main thread
